requirements.txt
asgiref==3.3.1
Django==3.1.3
django-autoslug==1.9.8
django-crispy-forms==1.12.0
django-multiselectfield==0.1.12
django-site-configs==0.1.0
django-user-agents==0.4.0
pytz==2020.4
sqlparse==0.4.1
ua-parser==0.10.0
user-agents==2.2.0
